[enh] Added iterative IDSchema.validate implementation and DoS Protection
The iterative version is now the default one. The recursive version can still be
used by passing `iterative=False` to IDSchema.validate.
Additionally a DoS Protection is implemented to restrict the maximum number of
reductions in the algorithm. This number can be set with the keyword argument
`max_reductions`.
Also added tests for all new features.